create and display initial group code

+++ src/main/kotlin/eu/asquare/droplets/data/Group.kt
@@ -11,7 +11,7 @@ class Group(
     val id: Long,
 
     @Column(columnDefinition = "SMALLINT UNSIGNED")
-    val code: Short,
+    val code: Int,
 
     @OneToMany(mappedBy = "group")
     val users: Set<User> = setOf()
